ServerUtility/requirements.txt
pika
psutil
pyinstaller
python-socketio
requests
packaging
aiohttp